🧹 Extract animation and keyboard logic out of ChartContainer.tsx
